Add routing management feature
New routing management feature allows dnsconfd to submit routes
to used DNS servers when necessary. This means situtation
when only one interface is able and allowed to contact the server.
Submitting of routes is done through Network Manager as it is
so far the default supported environment.
Feature can be turned off by setting handle-routing to no.